changed structure and indention
